pyproject.toml
[project]
name = "medical-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "ipykernel>=6.29.5",
    "langchain>=0.3.25",
    "langchain-community>=0.3.23",
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=2.1.4",
    "langchain-openai>=0.3.16",
    "langchain-pinecone>=0.2.5",
    "openai>=1.77.0",
    "pinecone[grpc]>=6.0.2",
    "pypdf>=5.4.0",
    "python-dotenv>=1.1.0",
    "sentence-transformers>=4.1.0",
]
